# openapi.yml
openapi: 3.0.3
info:
  title: English Learning App API
  description: API for managing users and audio files in the English learning application
  version: 1.2.1
  contact:
    name: English Learning App
    email: support@englishlearningapp.com
servers:
  - url: https://1bvgxdmq64.execute-api.us-east-1.amazonaws.com/dev
    description: Development server
security:
  - CognitoUserPool: []
paths:
...
  /users:
    get:
      operationId: getAllUsers
      summary: Get all users (admin only)
      tags: [Users]
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createUserProfile
      summary: Create user profile from JWT claims
      tags: [Users]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User profile created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}:
    get:
      operationId: getUserById
      summary: Get user by ID
      tags: [Users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User UUID
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - can only access own profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateUser
      summary: Update user profile
      tags: [Users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - can only update own profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteUserProfile
      summary: Delete user profile (local only)
      tags: [Users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User UUID
      responses:
        '204':
          description: User profile deleted successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - can only delete own profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/change-password:
    post:
      operationId: changePassword
      summary: Change user password
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized or invalid current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/signup:
    post:
      operationId: signUpUser
      summary: Sign up new user (public endpoint)
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: User signed up successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUpResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/confirm:
    post:
      operationId: confirmSignUp
      summary: Confirm user sign up (public endpoint)
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmSignUpRequest'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad request or invalid confirmation code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    CognitoUserPool:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito User Pool JWT token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Success message
      required: [message]
...
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email (optional, extracted from JWT if not provided)
        name:
          type: string
          description: User display name (optional, extracted from JWT if not provided)
      additionalProperties: false
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User UUID from Cognito
        email:
          type: string
          format: email
          description: User email address
        name:
          type: string
          description: User display name
        createdAt:
          type: string
          format: date-time
          description: Profile creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Profile last update timestamp
      required: [id, email, name, createdAt, updatedAt]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          description: Updated display name
        email:
          type: string
          format: email
          description: Updated email address (must match Cognito)
      additionalProperties: false
    ChangePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
          description: Current password
          minLength: 8
        newPassword:
          type: string
          description: New password
          minLength: 8
      required: [oldPassword, newPassword]
      additionalProperties: false

    # Auth Schemas
    SignUpRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          description: User password
          minLength: 8
        name:
          type: string
          description: User display name (optional)
      required: [email, password]
      additionalProperties: false
    SignUpResponse:
      type: object
      properties:
        message:
          type: string
          example: User signed up successfully. Please check your email for verification code.
        userSub:
          type: string
          description: Cognito user UUID
        codeDeliveryDetails:
          type: object
          properties:
            attributeName:
              type: string
              example: email
            deliveryMedium:
              type: string
              example: EMAIL
            destination:
              type: string
              example: t***@example.com
      required: [message, userSub]
    ConfirmSignUpRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email address
        confirmationCode:
          type: string
          description: Verification code sent to email
          pattern: '^[0-9]{6}$'
      required: [email, confirmationCode]
      additionalProperties: false
...